steelcity-ops-analytics/requirements.txt
pandas
numpy
SQLAlchemy
python-dotenv
scikit-learn
streamlit
plotly
psycopg2-binary
